clean up test-lib modules and related
