+++ marconi-chain-index/test-lib/Test/Integration.hs
@@ -1,9 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE NumericUnderscores #-}
-{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TupleSections #-}
-{-# OPTIONS_GHC -fplugin-opt PlutusTx.Plugin:target-version=1.0.0 #-}
 
 -- | Utilities for integration tests using tools from the `cardano-node-emulator` project.
 module Test.Integration (
@@ -15,35 +12,19 @@ module Test.Integration (
 import Cardano.Api.Extended.IPC qualified as C
 import Cardano.Api.Shelley qualified as C
 import Cardano.BM.Data.LogItem (LOContent (LogMessage), loContent)
-import Cardano.BM.Tracing qualified as BM
 import Cardano.Node.Emulator.Generators (knownAddresses, knownXPrvs)
 import Cardano.Node.Emulator.Internal.Node.Chain (ChainEvent (SlotAdd))
 import Cardano.Node.Emulator.Internal.Node.TimeSlot qualified as E.TimeSlot
 import Cardano.Node.Emulator.LogMessages (EmulatorMsg (ChainEvent))
 import Cardano.Node.Socket.Emulator qualified as E
 import Cardano.Node.Socket.Emulator.Types qualified as E.Types
-import Control.Concurrent (threadDelay)
-import Control.Monad (replicateM)
-import Control.Monad.Error.Class (MonadError)
-import Control.Monad.IO.Class (MonadIO, liftIO)
+import Control.Monad.IO.Class (MonadIO)
 import Control.Tracer (condTracing)
-import Data.Bifunctor (first)
 import Data.Default (def)
-import Data.Map qualified as Map
 import Data.Maybe (listToMaybe)
 import Data.Time.Clock.POSIX (getPOSIXTime)
-import Data.Word (Word64)
-import Hedgehog qualified as H
-import Hedgehog.Extras.Test.Base qualified as H
-import Hedgehog.Gen qualified as H.Gen
-import Hedgehog.Range qualified as Range
 import Ledger.Test (testnet)
-import Marconi.ChainIndex.Types (SecurityParam)
-import Marconi.Core qualified as Core
-import PlutusLedgerApi.V1 qualified as PlutusV1
-import PlutusLedgerApi.V2 qualified as PlutusV2
-import PlutusTx qualified
-import System.Exit (exitFailure)
+import Test.Gen.Marconi.ChainIndex.MintTokenEvent qualified as Gen.MintTokenEvent
 import Test.Helpers qualified as Helpers
 
 -- TODO: PLT-8098 need to shut down the node emulator properly when the test is done.
@@ -146,25 +127,19 @@ mkValidatedTxBodyWithFee
   -> C.TxBodyContent C.BuildTx C.BabbageEra
   -> C.Lovelace
   -- ^ Lovelace to fix in in TxOut, before applying fees
-  -> Word
+  -> Int
   -- ^ The number of Shelley key witnesses
   -> Either C.TxBodyError (C.TxBody C.BabbageEra)
 mkValidatedTxBodyWithFee ledgerPP networkid address txbodyc lovelace nKeywitnesses =
   C.createAndValidateTransactionBody txbodyc
     >>= C.createAndValidateTransactionBody . updateTxWithFee txbodyc
   where
-    -- TODO: PLT-8098 need to get address from txbody or somewhere
-    -- Take the txMintValue and add the fee-adjusted lovace to build a single TxOut.
     rebuildTxOutWithFee
       :: C.TxBodyContent C.BuildTx C.BabbageEra -> C.Lovelace -> C.TxOut C.CtxTx C.BabbageEra
     rebuildTxOutWithFee txbodyc' fee =
       Helpers.mkTxOut address $
-        C.lovelaceToValue (lovelace - fee) <> Helpers.getValueFromTxMintValue (C.txMintValue txbodyc')
-    -- TODO: PLT-8098 this consistently produces a "fee too small" error. the fee is off by < 1k
-    -- It uses cardano-ledger's transaction fee calculation function. does that differ from node
-    -- emulators?
-    -- calculateFee :: C.TxBody C.BabbageEra -> C.Lovelace
-    -- calculateFee txbody' = C.evaluateTransactionFee (C.unLedgerProtocolParameters ledgerPP) txbody' nKeywitnesses 0
+        C.lovelaceToValue (lovelace - fee)
+          <> Gen.MintTokenEvent.getValueFromTxMintValue (C.txMintValue txbodyc')
     updateTxWithFee
       :: C.TxBodyContent C.BuildTx C.BabbageEra
       -> C.TxBody C.BabbageEra
@@ -173,7 +148,7 @@ mkValidatedTxBodyWithFee ledgerPP networkid address txbodyc lovelace nKeywitness
       C.setTxOuts [rebuildTxOutWithFee txbodyc' fee] $
         C.setTxFee (C.TxFeeExplicit C.TxFeesExplicitInBabbageEra fee) txbodyc'
       where
-        fee = calculateFee ledgerPP (length $ C.txIns txbodyc') 1 0 1 networkid txbody'
+        fee = calculateFee ledgerPP (length $ C.txIns txbodyc') 1 0 nKeywitnesses networkid txbody'
 
 mkUnbalancedTxBodyContentFromTxMintValue
   :: (C.TxValidityLowerBound C.BabbageEra, C.TxValidityUpperBound C.BabbageEra)
@@ -185,32 +160,14 @@ mkUnbalancedTxBodyContentFromTxMintValue
 mkUnbalancedTxBodyContentFromTxMintValue validityRange ledgerPP address txIns txMintValue =
   (Helpers.emptyTxBodyContent validityRange ledgerPP)
     { C.txIns = map (,C.BuildTxWith $ C.KeyWitness C.KeyWitnessForSpending) txIns
-    , C.txOuts = [Helpers.mkTxOut address $ Helpers.getValueFromTxMintValue txMintValue]
+    , C.txOuts = [Helpers.mkTxOut address $ Gen.MintTokenEvent.getValueFromTxMintValue txMintValue]
     , C.txMintValue = txMintValue
     , C.txInsCollateral = C.TxInsCollateral C.CollateralInBabbageEra txIns
     }
 
 {- Indexers and queries -}
--- H.failMessage GHC.callStack
 
--- | TODO: PLT-8098 blocks until you get one. use the MarconiTrace and provide a better message.
-queryFirstResultWithRetry :: Int -> Word64 -> H.Integration [a] -> H.Integration a
-queryFirstResultWithRetry ntries _ _
-  | ntries <= 0 =
-      liftIO $
-        putStrLn "Max integration test query tries reached" >> exitFailure
-queryFirstResultWithRetry ntries delay action = do
-  res <- listToMaybe <$> action
-  case res of
-    Nothing ->
-      liftIO (threadDelay $ fromIntegral delay)
-        >> queryFirstResultWithRetry (ntries - 1) delay action
-    Just x -> pure x
-
--- TODO: PLT-8098 confirm that txOuts are not used for fee calculation and modify this
--- and or mkValidatedTxBodyWithFee.
-
--- TODO: PLT-8098 taken from legacy. review and remove if not needed
+-- | Wrapper for @C.'estimateTransactionFee'@.
 calculateFee
   :: (C.IsShelleyBasedEra era)
   => C.LedgerProtocolParameters era
@@ -230,12 +187,11 @@ calculateFee ledgerPP nInputs nOutputs nByronKeyWitnesses nShelleyKeyWitnesses n
         (C.makeSignedTransaction [] txBody)
         nInputs
         nOutputs
-        -- TODO: PLT-8098 note the swap
         nShelleyKeyWitnesses
         nByronKeyWitnesses
 
-{- | TODO: PLT-8098 sign the validated tx body and submit to node emulator within H.Integration
-context, and wait for it to be submitted over the protocol.
+{- | Sign the validated tx body, submit to node emulator
+ and wait for it to be submitted over the protocol.
 -}
 signAndSubmitTx
   :: (MonadIO m)
@@ -246,58 +202,3 @@ signAndSubmitTx
 signAndSubmitTx info witnesses txbody = Helpers.submitAwaitTx info (tx, txbody)
   where
     tx = C.makeSignedTransaction witnesses txbody
-
--- TODO: PLT-8098 All below copied from legacy. Reevaluate if needed.
--- They should go in a different module.
-
--- | TODO: PLT-8098
-genTxMintValue :: H.Gen (C.TxMintValue C.BuildTx C.BabbageEra)
-genTxMintValue = genTxMintValueRange 1 100
-
-genTxMintValueRange :: Integer -> Integer -> H.Gen (C.TxMintValue C.BuildTx C.BabbageEra)
-genTxMintValueRange min' max' = do
-  n :: Int <- H.Gen.integral (Range.constant 1 5)
-  policyAssets <- replicateM n genAsset
-  let (policyId, policyWitness, mintedValues) = mkMintValue commonMintingPolicy policyAssets
-      buildInfo = C.BuildTxWith $ Map.singleton policyId policyWitness
-  pure $ C.TxMintValue C.MultiAssetInBabbageEra mintedValues buildInfo
-  where
-    genAsset :: H.Gen (C.AssetName, C.Quantity)
-    genAsset = (,) <$> genAssetName <*> genQuantity
-    genQuantity = C.Quantity <$> H.Gen.integral (Range.constant min' max')
-
-genAssetName :: H.Gen C.AssetName
-genAssetName = C.AssetName <$> H.Gen.bytes (Range.constant 1 5)
-
-mkMintValue
-  :: MintingPolicy
-  -> [(C.AssetName, C.Quantity)]
-  -> (C.PolicyId, C.ScriptWitness C.WitCtxMint C.BabbageEra, C.Value)
-mkMintValue policy policyAssets = (policyId, policyWitness, mintedValues)
-  where
-    serialisedPolicyScript :: C.PlutusScript C.PlutusScriptV1
-    serialisedPolicyScript = C.PlutusScriptSerialised $ PlutusV2.serialiseCompiledCode policy
-
-    policyId :: C.PolicyId
-    policyId = C.scriptPolicyId $ C.PlutusScript C.PlutusScriptV1 serialisedPolicyScript :: C.PolicyId
-
-    executionUnits :: C.ExecutionUnits
-    executionUnits = C.ExecutionUnits{C.executionSteps = 300000, C.executionMemory = 1000}
-    redeemer :: C.ScriptRedeemer
-    redeemer = C.unsafeHashableScriptData $ C.fromPlutusData $ PlutusV1.toData ()
-    policyWitness :: C.ScriptWitness C.WitCtxMint C.BabbageEra
-    policyWitness =
-      C.PlutusScriptWitness
-        C.PlutusScriptV1InBabbage
-        C.PlutusScriptV1
-        (C.PScript serialisedPolicyScript)
-        C.NoScriptDatumForMint
-        redeemer
-        executionUnits
-
-    mintedValues :: C.Value
-    mintedValues = C.valueFromList $ map (first (C.AssetId policyId)) policyAssets
-
-type MintingPolicy = PlutusTx.CompiledCode (PlutusTx.BuiltinData -> PlutusTx.BuiltinData -> ())
-commonMintingPolicy :: MintingPolicy
-commonMintingPolicy = $$(PlutusTx.compile [||\_ _ -> ()||])
